+++ src/main/java/com/flawden/TaskManagerAPI/service/impl/TaskServiceImpl.java
@@ -17,7 +17,7 @@ public class TaskServiceImpl implements TaskService {
     }
 
     @Override
-    public List<Task> getTasksWithPagination(int page) {
+    public List<Task> getTasksWithPagination(Long page) {
         return List.of();
     }
